requirements.txt
gradio
huggingface_hub
pycryptodome
datasets
pyyaml